axf/AXF_VUE/requirements.txt
alembic==1.0.10
amqp==2.5.0
aniso8601==6.0.0
asn1crypto==0.24.0
billiard==3.5.0.5
celery==4.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
Django==1.11.7
django-celery-results==1.0.4
django-cors-headers==3.0.2
django-debug-toolbar==1.10.1
django-redis==4.9.0
django-redis-cache==2.0.0
django-tinymce==2.7.0
djangorestframework==3.9.4
drf-extensions==0.5.0
Flask==1.0.2
Flask-Caching==1.7.1
Flask-Migrate==2.5.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.1.0
Mako==1.0.10
MarkupSafe==1.1.1
Pillow==5.2.0
pycparser==2.18
pycryptodomex==3.7.2
PyMySQL==0.9.2
python-alipay-sdk==1.10.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
SQLAlchemy==1.3.3
sqlparse==0.2.4
urllib3==1.23
uWSGI==2.0.17.1
vine==1.3.0
Werkzeug==0.15.2
